Resolved comment
